/follower
- /follower : check if token follow user_id

+++ controllers/follower.controller.js
@@ -1,43 +1,64 @@
 const follower = require('../models/follower.model');
 
 module.exports.follow = async (req, res) => {
-    if(req.user_data){
-        if(!req.follower_data){
-            link = {
-                userA : req.user._id,
-                userB : req.body.user_id,
+    if(req.user){
+        if(req.user_data){
+            if(!req.follower_data){
+                link = {
+                    userA : req.user._id,
+                    userB : req.body.user_id,
+                }
+                const newFollower = new follower(link);
+                await newFollower.save();
+                res.status(200).json("Follow : relationship created successfully");
             }
-            const newFollower = new follower(link);
-            await newFollower.save();
-            res.status(200).json("Follow : relationship created successfully");
+            else {
+                res.status(400).json("Follow : relationship already exists");
+            }
         }
-        else {
-            res.status(400).json("Follow : relationship already exists");
-        }
-    }
-    else{
-        res.status(400).json("Follow : user don't exist");
-    }    
+        else{
+            res.status(400).json("Follow : user don't exist");
+        }   
+    } 
 };
 
 module.exports.unfollow = async (req, res) => {
-    if(req.user_data){
-        if(req.follower_data){
-            link = {
-                userA : req.user._id,
-                userB : req.body.user_id,
+    if(req.user){
+        if(req.user_data){
+            if(req.follower_data){
+                link = {
+                    userA : req.user._id,
+                    userB : req.body.user_id,
+                }
+                await follower.deleteMany(link);
+                res.status(200).json("Unfollow: relationship deleted successfully");
             }
-            await follower.deleteMany(link);
-            res.status(200).json("Unfollow: relationship deleted successfully");
+            else {
+                res.status(400).json("UnFollow : relationship don't exists");
+            }
         }
-        else {
-            res.status(400).json("UnFollow : relationship don't exists");
-        }
+        else{
+            res.status(400).json("UnFollow : user don't exist");
+        }    
     }
-    else{
-        res.status(400).json("UnFollow : user don't exist");
-    }    
 };  
+module.exports.doIfollow = async (req, res) => {
+    if(req.user){
+        if(req.user_data){
+            if(req.follower_data){
+                res.status(200).json("Followed");
+            }
+            else {
+                res.status(200).json("UnFollowed");
+            }
+        }
+        else{
+            res.status(400).json("UnFollow : user don't exist");
+        }  
+    }  
+};  
+
+
 
 //                              //
 //-------- MiddleWares----------//
